+++ packages/sessions/src/utils/simple-logger.ts
@@ -13,18 +13,22 @@ export class SimpleLoggerImpl implements Logger {
     }
 
     debug(message: string, ...args: any[]): void {
+        // eslint-disable-next-line no-console
         console.debug(`[${this.prefix}:DEBUG] ${message}`, ...args);
     }
 
     info(message: string, ...args: any[]): void {
+        // eslint-disable-next-line no-console
         console.info(`[${this.prefix}:INFO] ${message}`, ...args);
     }
 
     warn(message: string, ...args: any[]): void {
+        // eslint-disable-next-line no-console
         console.warn(`[${this.prefix}:WARN] ${message}`, ...args);
     }
 
     error(message: string, ...args: any[]): void {
+        // eslint-disable-next-line no-console
         console.error(`[${this.prefix}:ERROR] ${message}`, ...args);
     }
 }